better MVC
