+++ src/main/java/com/KUAlchemists/backend/handlers/BoardHandler.java
@@ -2,6 +2,8 @@ package com.KUAlchemists.backend.handlers;
 
 import com.KUAlchemists.backend.engine.GameEngine;
 
+import java.util.ArrayList;
+
 public class BoardHandler {
 
 
@@ -17,7 +19,12 @@ public class BoardHandler {
         return INSTANCE;
     }
 
-    public void endTheTour() {
-        GameEngine.getInstance().nextTour();
+    /**
+     * This method is called when a player wants to end the tour.
+     * @return The next tour.
+     * @see GameEngine#nextTour()
+     */
+    public ArrayList<Integer> endTheTour() {
+        return GameEngine.getInstance().nextTour();
     }
 }
